extract applyFilters into a separate function
